+++ e2e/starter-e2e/src/generators.spec.ts
@@ -52,9 +52,11 @@ describe('Generators tests', () => {
         it('exists lint files', async () => {
           const commitlintrc = await exists(join(genFolder, projectFolder, '.commitlintrc.cjs'));
           const eslint = await exists(join(genFolder, projectFolder, 'eslint.config.ts'));
+          const eslintflatignore = await exists(join(genFolder, projectFolder, '.eslintflatignore'));
           const stylelintrc = await exists(join(genFolder, projectFolder, '.stylelintrc.cjs'));
           expect(commitlintrc).toBeTruthy();
           expect(eslint).toBeTruthy();
+          expect(eslintflatignore).toBeTruthy();
           expect(stylelintrc).toBeTruthy();
         });